Converts a bunch more nodeunit tests to mocha
